remove more explicit cluster
